+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 * Adding some check for sql query result
 * Adding class to check configuration (Database extension and structure, Project layers required)
+* Improve the import script example for French IGN BdTopo
+* Spatial query performance improvements
 
 ## 0.1.0 - 2021-06-22